move utils to proper place
