+++ cumulus/zombienet/zombienet-sdk-helpers/Cargo.toml
@@ -16,4 +16,6 @@ cumulus-primitives-core = { workspace = true, default-features = true }
 tokio = { workspace = true, features = ["rt-multi-thread", "macros", "time"] }
 zombienet-sdk = { workspace = true }
 futures = { workspace = true }
+zombienet-configuration = { workspace = true }
+
