+++ cumulus/zombienet/zombienet-sdk/tests/utils.rs
@@ -1,16 +1,7 @@
 // Copyright (C) Parity Technologies (UK) Ltd.
 // SPDX-License-Identifier: Apache-2.0
 
-use cumulus_zombienet_sdk_helpers::{
-	create_assign_core_call, submit_extrinsic_and_wait_for_finalization_success_with_timeout,
-};
-use zombienet_sdk::{
-	subxt::{OnlineClient, PolkadotConfig},
-	tx_helper::{ChainUpgrade, RuntimeUpgradeOptions},
-	LocalFileSystem, Network, NetworkConfig, NetworkNode,
-};
-
-use zombienet_configuration::types::AssetLocation;
+use zombienet_sdk::{LocalFileSystem, Network, NetworkConfig};
 
 pub const BEST_BLOCK_METRIC: &str = "block_height{status=\"best\"}";
 
@@ -28,41 +19,4 @@ pub async fn initialize_network(
 	network.detach().await;
 
 	Ok(network)
-}
-
-pub async fn runtime_upgrade(
-	network: &Network<LocalFileSystem>,
-	node: &NetworkNode,
-	para_id: u32,
-	wasm_path: &str,
-) -> Result<(), anyhow::Error> {
-	log::info!("Performing runtime upgrade for parachain {}, wasm: {}", para_id, wasm_path);
-	let para = network.parachain(para_id).unwrap();
-
-	para.perform_runtime_upgrade(node, RuntimeUpgradeOptions::new(AssetLocation::from(wasm_path)))
-		.await
-}
-
-pub async fn assign_cores(
-	node: &NetworkNode,
-	para_id: u32,
-	cores: Vec<u32>,
-) -> Result<(), anyhow::Error> {
-	log::info!("Assigning {:?} cores to parachain {}", cores, para_id);
-
-	let assign_cores_call =
-		create_assign_core_call(&cores.into_iter().map(|core| (core, para_id)).collect::<Vec<_>>());
-
-	let client: OnlineClient<PolkadotConfig> = node.wait_client().await?;
-	let res = submit_extrinsic_and_wait_for_finalization_success_with_timeout(
-		&client,
-		&assign_cores_call,
-		&zombienet_sdk::subxt_signer::sr25519::dev::alice(),
-		60u64,
-	)
-	.await;
-	assert!(res.is_ok(), "Extrinsic failed to finalize: {:?}", res.unwrap_err());
-	log::info!("Cores assigned to the parachain");
-
-	Ok(())
 }
